requirements.txt
flask>=3.0
psycopg2-binary>=2.9
pymysql>=1.1
python-dotenv>=1.0
openpyxl>=3.1
gunicorn>=21.0